requirements.txt
streamlit
google-generativeai
folium
streamlit-folium
geopy
pandas